--- src/Components/test/E2ETestMigration/Infrastructure/AssemblyInfo.AssemblyFixtures.cs
@@ -1,9 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using Microsoft.AspNetCore.E2ETesting;
-using Microsoft.AspNetCore.Testing;
-using Xunit;
-
-[assembly: AssemblyFixture(typeof(SeleniumStandaloneServer))]
-[assembly: AssemblyFixture(typeof(SauceConnectServer))]
